+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 9.4.0
 - PPI-807 - Improved script loading performance in the Storefront
 - PPI-924 - Added more explicit error messages for errors happening during express checkout
+- PPI-950 - Changed the shipment carrier field to a dropdown in the Administration
 - PPI-951 - Fixes an issue, where only generic error messages were displayed during checkout without Smart Payment Buttons
 - PPI-953 - Fixes an issue, where shipping tracking codes were not synced with an invalid carrier